+++ Core/Driver/DriverBinding/Interrupt/Input.cpp
@@ -38,7 +38,7 @@ namespace Driver
         debug("Searching for conflicting drivers...");
         if (fexExtended->Driver.OverrideOnConflict)
         {
-            std::vector<uint64_t> DriversToRemove = std::vector<uint64_t>();
+            std::vector<size_t> DriversToRemove = std::vector<size_t>();
             foreach (auto Drv in Drivers)
             {
                 FexExtended *fe = ((FexExtended *)((uintptr_t)Drv.Address + EXTENDED_SECTION_ADDRESS));
