+++ Core/PeripheralComponentInterconnect.cpp
@@ -789,7 +789,7 @@ namespace PCI
     }
 #endif
 
-    void PCI::EnumerateFunction(uintptr_t DeviceAddress, uint64_t Function)
+    void PCI::EnumerateFunction(uintptr_t DeviceAddress, uintptr_t Function)
     {
         uintptr_t Offset = Function << 12;
         uintptr_t FunctionAddress = DeviceAddress + Offset;
@@ -805,7 +805,7 @@ namespace PCI
 #endif
     }
 
-    void PCI::EnumerateDevice(uintptr_t BusAddress, uint64_t Device)
+    void PCI::EnumerateDevice(uintptr_t BusAddress, uintptr_t Device)
     {
         uintptr_t Offset = Device << 15;
         uintptr_t DeviceAddress = BusAddress + Offset;
@@ -819,7 +819,7 @@ namespace PCI
             EnumerateFunction(DeviceAddress, Function);
     }
 
-    void PCI::EnumerateBus(uintptr_t BaseAddress, uint64_t Bus)
+    void PCI::EnumerateBus(uintptr_t BaseAddress, uintptr_t Bus)
     {
         uintptr_t Offset = Bus << 20;
         uintptr_t BusAddress = BaseAddress + Offset;
